# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Audio Guide 2026 API
  version: 1.13.0
  description: Canonical API for Audio Guide 2026. Includes fail-closed caching rules.
servers:
  - url: https://audiogid-api.vercel.app/v1
    description: Production Server
paths:
...
  /admin/media/presign:
    post:
      tags: [Admin, Media]
      summary: Presign Media Upload
      operationId: presignMedia
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PresignRequest' }
      responses:
        '200':
          description: Upload URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PresignResponse' }

  # --- Offline Bundles (ADR-013) ---
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    YookassaSignature:
      type: apiKey
      in: header
      name: X-Yookassa-Signature
  schemas:
...
    PresignRequest:
      type: object
      required: [filename, content_type]
      properties:
        filename: { type: string }
        content_type: { type: string }
        entity_type: { type: string, nullable: true }
        entity_id: { type: string, nullable: true }
    PresignResponse:
      type: object
      required: [upload_url, final_url]
      properties:
        upload_url: { type: string }
        final_url: { type: string }
        access: { type: string }
...